Rename label image validator
